+++ raiden/ui/web/src/app/components/join-dialog/join-dialog.component.ts
@@ -1,7 +1,7 @@
-import { Component, Inject } from '@angular/core';
-import { FormControl } from '@angular/forms';
+import { ChangeDetectorRef, Component, Inject, OnInit, ViewChild } from '@angular/core';
+import { FormBuilder } from '@angular/forms';
 import { MAT_DIALOG_DATA, MatDialogRef } from '@angular/material';
-import { BigNumber } from 'bignumber.js';
+import { TokenInputComponent } from '../token-input/token-input.component';
 
 export interface JoinDialogPayload {
     tokenAddress: string;
@@ -14,39 +14,34 @@ export interface JoinDialogPayload {
     templateUrl: './join-dialog.component.html',
     styleUrls: ['./join-dialog.component.css']
 })
-export class JoinDialogComponent {
+export class JoinDialogComponent implements OnInit {
 
-    private _decimals = 0;
-    public funds: FormControl = new FormControl(0);
+    @ViewChild(TokenInputComponent) tokenInput: TokenInputComponent;
+
+    form = this.fb.group({
+        amount: 0,
+        decimals: true
+    });
 
     constructor(
         @Inject(MAT_DIALOG_DATA) public data: JoinDialogPayload,
-        public dialogRef: MatDialogRef<JoinDialogComponent>
+        public dialogRef: MatDialogRef<JoinDialogComponent>,
+        private fb: FormBuilder,
+        private cdRef: ChangeDetectorRef,
     ) {
-        this._decimals = data.decimals;
+    }
+
+    ngOnInit(): void {
+        this.tokenInput.decimals = this.data.decimals;
+        this.cdRef.detectChanges();
     }
 
     public joinTokenNetwork() {
         const payload: JoinDialogPayload = {
             tokenAddress: this.data.tokenAddress,
-            funds: this.funds.value,
-            decimals: this._decimals
+            funds: this.tokenInput.tokenAmount.toNumber(),
+            decimals: this.tokenInput.tokenAmountDecimals
         };
         this.dialogRef.close(payload);
     }
-
-    public step(): string {
-        return (1 / (10 ** this._decimals)).toFixed(this._decimals).toString();
-    }
-
-    public decimals(): number {
-        return this._decimals;
-    }
-
-    public precise(value) {
-        if (value.type === 'input' && !value.inputType) {
-            this.funds.setValue(new BigNumber(value.target.value).toFixed(this._decimals));
-        }
-    }
-
 }